services/agent-a-service/pyproject.toml
[project]
name = "agent-a-service"
version = "0.1.0"
description = "Criteria extraction agent service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "shared",
    "inference",
    "events-py",
    "pymupdf>=1.24.0",
    "pymupdf4llm>=0.2.9",
    "diskcache>=5.6.3",
    "jinja2>=3.1.0",
    "langchain-google-vertexai>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/agent_a_service"]

[tool.ruff]
lint.select = ["E", "F", "W", "C", "N", "I", "D"]
lint.ignore = ["E203", "D203", "D213", "D100", "D413", "D104"]
lint.pydocstyle.convention = "google"
target-version = "py312"
line-length = 88
exclude = ["venv", ".venv", "docs"]

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "docs"]
plugins = []
mypy_path = ["../shared/src", "../inference/src", "../../libs/events-py/src"]

[tool.pytest.ini_options]
pythonpath = ["src", "../shared/src", "../inference/src", "../../libs/events-py/src"]
testpaths = ["tests"]
addopts = "--cov=agent_a_service --cov-branch --cov-report=term-missing"

[tool.uv.sources]
shared = { workspace = true }
inference = { workspace = true }
events-py = { workspace = true }
